refactor static error-code generation into factory
